py38 extra cleanup & testing (#88)
I realized I didn't merge the tests I had added, or some of the cleanup I did after adding the tests. The main change besides the test was:
(a) to invert the relation of the "is_x" checks and the "validate_x" checks so more appropriate errors could be used in each case
(b) take a method I had renamed to an "is_x" format such that it no longer raises an error by default, but returns a False like the name would suggest.