Make versions consistent (#101)
Last release, the contents of versions.py were inconsistent with the wheel that was produced. Fix that, and make sure we catch it in tests if they are inconsistent again. Unfortunately I can't think of a good (easy) way to have a "single source of truth" for the version that's accessible for the markdown, bazel build, and python module. So catching inconsistencies is likely the best option right now!