Add Resolutions table and data access APIs (#102)
Add a table to hold resolution objects, and add the DB and HTTP API endpoints to interact with it. This is a pre-req for putting K8s behind the API, because we need to put the docker image URI and user settings into the DB for the server to know how to tell k8s about them.

Note that this PR doesn't actually hook this table up to anything, it just creates the table and accessors. The next PR will hook it up.