Add `has_container_image` (#127)
With the bazel setup, you can do bazel run //foo:bar or bazel run //foo:bar_local, but it's really hard to set up a situation where an appropriate resolver gets used based on which of these you used. This PR adds an API that can be used for that purpose. I also used this as an opportunity to beef up our docs on how we integrate with Docker.