Persist resource reqs in DB (#123)
ResourceRequirements let users specify function-specific resource requirements. For example, a function can specify that it needs a particular type of Kubernetes node (e.g. GPU, high-mem).

At this time, these are used at runtime when launching K8s jobs, but they are not persisted in the DB. Persisting them is necessary to enable re-running pipelines from the UI and CLI, as well as to move kubernetes job launches behind the server.