Schedule resolution job behind the server (#144)
The last piece of getting k8s behind the server is to move the scheduling of the resolution job behind the server. This PR does that. Remaining work on this thread includes documenting how to deploy the server now that the server needs to have permissions to schedule things on k8s, but that will be left for a separate PR.