Remove env vars from resolution API response (#147)
The Resolutions table needs to hold env vars that should be mounted to the job containers for usage by those jobs. Resolutions also need to be readable--to read and transition their state, for example. But we don't want those env vars to be readable by anybody, due to potentially sensitive info being held in the env vars. This makes it so we scrub the env vars from the API reads.