Log uploads (#158)
We want to display logs in the UI; first we need to ingest them! This PR makes it so that we do that. A side-effect of this is that kubectl logs won't give you the full logs of an execution. However, I've made it so when you do kubectl logs <my pod> it will show you another kubectl command you can do to follow the logs.