Temporarily remove drop column from up migrations (#323)
This removes the drop column operations from the up direction DB
migrations in order to unblock users who are impacted by this issue:
#302.

The check that sqlite3 has a minimum version of 3.35.0 is also relaxed
to just warn instead of fail the starting of the server.

The effect is that users will be able to migrate to newer versions of
Sematic on both their local and cloud deployments even if their Python
installation did not come with sqlite3 >= 3.35.0, without requiring them
to install a newer version from sources.

This is a temporary measure until the issue is resolved in a more
sustainable way. The column that was previously removed will be kept in
the DB until we address the issue.

Tested by starting the server without a database file, and by migrating
down a database that contained relevant exception data, and back up
again.