Streamlining handling of artifacts in `LocalResolver` in order to fix a re-run bug (#332)
It turns out it is possible that certain input edges have neither an
`artifact_id` set nor a `source_run_id` set.

This is the case when the edge points to a value passed from outside the
parent future. For example:

```python
@sematic.func
def pipeline(a: int) -> int:
    f = foo()
    b = bar(f, a)
    return b
```
In this case, the edge for the `a` input to `bar` will have neither
`artifact_id` nor `source_run_id` set until `bar` is ready to be
scheduled. This prevents the graph from being cloned in the case where
an upstream run fails (e.g. `foo`).

This PR fixes this issue by creating artifacts as soon as there is a
concrete value for it, it does not wait for `bar` to be ready to get
scheduled.