Read logs ingested using "delta" format (#328)
This PR is the number 2 in a series. The series:
(1) (#306) Change log ingestion to write "deltas" containing only the
log lines since the last upload
(2) Change log reading to combine deltas to get the full lines
(3) Write a convenience script to combine logs from S3 to show logs for
us. So we can examine logs with only s3 access, without requiring access
to the UI

This PR's primary goal is to switch to reading to use log-line deltas
rather than full log files.

Testing
--------
Had a remote pipeline write 10k log lines to stdout and to stderr. See
[run](https://dev.sematic.cloud/pipelines/sematic.examples.bazel.pipeline.pipeline/6be6811d010d49b0ab6c511ff2c33a91),
the "add" run specifically. I then confirmed that infinite scroll and
"jump to end" worked--with and without a filter in use.