Fix Union serdes (#352)
Prior to this PR, `Union` had logic for serializing to json, but not for
deserializing from it. How could this have ever worked, you ask? Because
the serialization logic for `Union` just forwarded straight to the
serialization logic for the underlying type, and then on deserialization
the fallback generally fell back to something that "just worked" (either
pickle or assuming the object was already json encodable). However, in
some cases, this would fail. So I now added an actual
deserialize-from-json function for `Union`. It stores information about
which of the Unioned types the deserialization should ultimately use.
Note that this does NOT change the json summary logic, so the UI is not
impacted by this change.