Chance/refactor view redirection (#357)
Now we have a dedicated page to handle `pipelines/:calculatorPath`
routes. Therefore the process of delegating a sub-component to find the
correct run and redirect URL could be lifted to a higher level. This
saves some unnecessary communication across component layers.

This is also more consistent with the idea of deep linking, in that we
grant different URL patterns different semantics.
`pipelines/:calculatorPath` then has its own purpose and which does not
overlap with the semantic of `pipelines/:calculatorPath/:rootId`.

Co-authored-by: tscurtu <tudor@sematic.dev>