[bugfix] Cache version check (#378)
We have a mechanism to validate compatibility of version between client
and server. The check is supposed to be done once per interpreter
session and then be cached for subsequent calls. It is not cached.
This leads to every single API query being preceded by a version check.

This PR fixes this.