[bugfix] Missing broadcasts upon server-side run save (#379)
Every time the server saves a run, it needs to broadcast the fact that
the graph was updated. These two were missing leading to the DAG not
refreshing in the UI.

Ideally we would have a centralized way of doing this (so that any call
to `save_run` automatically broadcasts the update), but I don't want to
add a dependency to socketio in the `queries` module as it would mix
concerns.

Maybe we need an additional intermediate layer that takes care of
persisting objects (calling `queries` functions) and broadcasting to
socketio too.