[bugfix] Inline future status were not being updated by `CloudResolver` prior to executing (#380)
With `CloudResolver`, the server is responsible for starting k8s jobs,
and also updating the run status to `SCHEDULED`.

Inline runs are not scheduled by the server, therefore the usual logic
needs to be applied.

This bug lead to inline runs remaining in the `CREATED` state until
completion. This would have been caught by #107.