Enable Sematic context (#377)
#138

Introduces a new context object that can be used to obtain the ids of
the active run/root run as well as the Resolver class. Example:

```python
import sematic

@sematic.func
def my_func() -> None:
    print("Root id is: {}".format(sematic.context().root_id))
    return None
```

This PR serves 3 purposes. The first two are related to the start of
Spark work, the last is general.
1. In order to have behavior that differs based on which resolver is
being used (ex: allocate local or remote spark cluster), we will need a
context object like this
2. In order to allow the resolver logic to have a "hook" into how
references to external resources (ex: a spark cluster) are registered,
we will need such a context. Silent resolver can store references to
these resources in memory, while the others store in a DB.
3. there are some legit use cases where It is useful to be able to get
the id of the current run/root run (see #138)

Testing
--------
In addition to unit tests, also executed a cloud run in which I printed
the context from within a function. Ensured that the logs showed the
correct context. Run
[here](https://dev1.dev-usw2-sematic0.sematic.cloud/pipelines/sematic.examples.bazel.pipeline.pipeline/743c03cd549c4dccaab1be55acc507ca).
View source code and logs for `add` to see the result.