Gets rid of passing down `changeRootRunId` (#367)
Description:

There is no need to have `changeRootRunId` passed down from
`<PipelineRunView />,` through `<PipelineBar />,` then through
`<NotesPanel />,` into `<NoteView />` (or similarly, from
`<PipelineRunView />` into `<PipelineBar>`). If `<NoteView />` wants to
change the root run, the best way is to make URL navigation and render a
new page fresh. This doesn't have much performance drain because if a
node is not changed on the virtual DOM tree, it will persist and
potentially not be re-rendered. In the meantime, the code flow logic
could be simplified.

This idea also works well with the future's comprehensive deep-linking
implementation.