Fix log continuation bug (#394)
Fixes a bug where you could get the following behavior:

- run is in progress, you scroll to the end
- you request more logs
- you get back more logs from the beginning

The issue was this:
- the code for constructing the cursor to be returned uses the place
where it stopped searching
- if you requested more lines and there were 0 more available, there was
no information available about where the code had stopped searching

The fix was to have the code return the same cursor position as it was
supposed to start from if it didn't find any log lines (but might still
find more later because the run is still running).

Testing
--------
- Added a unit test to emulate this case. Confirmed the test fails with
the old code, and passes with the new.
- Tested using `bazel run //sematic/examples/testing_pipeline:__main__
-- --sleep 1000 --cloud --detach` and following the logs on the
`do_sleep` func. Used the UI from [this
branch](https://github.com/sematic-ai/sematic/pull/391) to confirm that
when I asked for more logs after it had reached the end of the
available, it gave me the correct next logs.