Makes RunID links more explicit visually (#397)
Currently, the links on the notes are not evident enough to be seen as a
link.

This change makes the links more evident.

From 
<img width="312" alt="image"
src="https://user-images.githubusercontent.com/1046489/207132004-daec1235-f670-4c7d-bdae-49293a820c99.png">

To:
<img width="321" alt="image"
src="https://user-images.githubusercontent.com/1046489/207131932-28569029-e49c-4e0c-9184-d3135f1fe5ef.png">


It used to display as plain text instead of the link because we
implemented it that way. Now when we use `<link>,` it will behave more
naturally like an HTML native link. We gain the benefit of showing
visited states of the link.