Add back worker url override (#399)
There was a feature that would allow you to configure the workers to use
a different URL to talk to the server than the end-users did. This is
useful if your networking setup is such that user machines will hit some
sort of proxy that can't be used in-cluster.

The old setting was left in as a server setting, and the docs still
reference it. It just wasn't hooked up to anything. So this PR just goes
the last mile to hook it back up--this time functioning fully
server-side rather than relying on the configuration to be set by the
users. Luckily, in-cluster, you can just use the k8s built-in DNS to
refer to the API via its k8s service name. This means that for helm
deployments, we can set this config automatically.

Testing
-------
Deployed this code to dev1 using the helm chart. Then for my client
code, modified the API client to print out the URLs it constructed.
Executed the testing pipeline, confirmed the URLs were using the k8s
service name, and successfully. See
[run](https://dev1.dev-usw2-sematic0.sematic.cloud/pipelines/sematic.examples.testing_pipeline.pipeline.testing_pipeline/90d5147363e24910bb85a24aa7e5641c)
and its logs.