[1/2] Reactors the state management of Pipeline scenarios (#410)
Elevate `rootRun`, `resolution`, and `selectedPanelItem` into supervisor
contexts

Maintain the selected nested run ( `selectedRuns`) in a supervisor
context. This is to pave the way to implement deep linking more easily.

What's next:
Due to the size of this PR, a follow-up PR will divide RunPanels into
GraphPanel and RunDetailsPanel sections. So we have a dedicated
component to deal with Graph manipulations.
Graph data derivative runsById data will also stay in a context.