Avoid Artifact overwrite (#420)
Artifacts are addressed by content, with the ID being generated from the
type and value themselves. When a Resolution generates an Artifact which
has been seen before and which is present in the DB, this DB entry is
completely overwritten.

The consequence given the current array of Artifact fields is that the
`created_at` field is always updated to the latest Resolution time of
any pipeline that produces it, meaning it can't be actually trusted.

This PR changes the behavior to not update artifacts that already exist.

## Testing

Added 2 unit tests:
- validate creating a new Artifact
- validate updating an existing Artifact does not overwrite the
`created_at` column.