Raise an error if there are unused futures (#421)
Closes #384

A very common source of confusion for users is when they "call" a func
inside of another one, but don't have any data dependency of the final
result on the output of the inner func. This PR makes this into an error
with a helpful message, as well as beefed up docs on this circumstance.