Move(rename) some UI files for better organization.  (#429)
This PR moves files into specific folders for better organizing them. 

Background: Some files under `component` folders are not granular enough
to be reusable as **components**. So they are moved to `pipeline` folder
to denote that they belong to the business of the ”pipeline page“.

No files are renamed, but files are moved to a different location. 

Q: Why are you doing this PR now?
A: It is the interim of different large-scale refactorings. I didn't
want to do it alongside the previous refactoring because that would
pollute the Git diff by eclipsing the real changes since the moved files
would become completely new files and all their lines are new changes.
It would cause troubles for code reviews.