Update make_artifact to use value type for Union (#342)
This PR fixes issue #235, where we display the json summary for the list
instead of the actual contents of the list. This code will work well for
Optional Types, but it's not perfect.

Suppose the user defines a type as `typing.Union[typing.List[int],
typing.Tuple[int]]` or `typing.Union[int, float]` etc, we will not get
the correct output. However, to fix this we would need to check the type
of the value summary and choose the view based on it. I am hesitant to
do this because checking types in Typescript for objects that were
defined in Python seems weird to me at least.

The ideal solution would be to figure out the value type and return the
correct type from the backend instead of the `Union` type. I would like
to get your views on whether I should go for this backend fix instead.
Thanks.