Implement deep link for selected runs (#433)
Implements deep link thanks to the help of [Jotai](https://jotai.org/)
library.

Deep link has a pattern of 
```
http://localhost:3000/pipelines/sematic.examples.add.pipeline.pipeline/46b01356e2864c4f917d1d9125288ee7#selRun=%2224da63e8353d4cd3b49e0dbf96f881a4%22
```


Some other maintenance work:
* Changed the API for reporting selected runs to `setSelectedRunId`, so
we avoid client reporting with inconsistent runs. The real run instance
will be given by the `PipelinePanelsContext`, so this keeps the
references to the run instances as consistent as possible.