Add functionality to click artifacts in the DAG view (#426)
This PR addresses issue #331. I have added the code linking artifact
nodes in the DAG to the run panel.
Further, for input artifacts with no source run, I have added the
ability to scroll down to the relevant artifact automatically.

Only part missing in the PR, is to make the artifact green when it is
selected which I will address in the future.