[Bug] Fixes an issue that not being able to unmount a component causes unnecessary complexity in the downstream processing. (#444)
Fixes an issue that switching runs from the dropdown box does not switch
the pages.

The reason why it didn't work is that the run data is not updated in
time when we make a judgement whether the current run is a nested run in
:


https://github.com/sematic-ai/sematic/blob/8e9ae391da56d1c2f2c4ec9617d1aee4b5d52761/sematic/ui/src/pipelines/PipelineRunView.tsx#L65-L73

In other words, this issue can also be fixed by
[507b53](https://github.com/sematic-ai/sematic/commit/9406f59bf66e7ef5fc33908207328fe743ae1930)

But I don't choose that way because there will be more complexities to
consider in the future if we don't unmount the component and still
attempt to fix things up based on the existing component state. IMHO,
unmounting and remounting the component fresh is worry-free and more
robust in the long run. Switching the run in the dropdown box will
trigger a significant re-rendering of the entire page data. We gain
little performance keeping the component live anyway.