Add props to cloned future to fix rerun name (#443)
This PR fixes issue #393 . I chose to hardcode the props for name and
tags because
https://github.com/sematic-ai/sematic/blob/main/sematic/db/models/factories.py#L30
erases the name and tags from the existing future. It seems like for
normal runs `.set` is called after this which sets the props but it
doesn't seem to be the case for rerun resolutions.

At least from my understanding, this is why this issue happens.