Prevents getting infinite loops with useEffect() (#469)
A bug has been discovered following the es lint refactoring in
https://github.com/sematic-ai/sematic/pull/451.

Previously, we wrongly relied on missing dependencies in the `useEffect`
dependency list as a feature. Adding the missing dependency causes a
component to render in a loop. Now we remove `authorsByEmail` as a state
that would trigger the re-rendering of the component and use
reference(`useRef`) instead (which doesn't trigger re-rendering, but can
still track the data).