Refactors: use context to pass state.  (#480)
Refactors to address the following two issues:

1. use context to pass the state of `selected run`, instead of passing
it through the component hierarchy. Also created a component for each
run details tab panel explicitly.

2. improved navigation logic to address an issue that redirection logic
creates unnecessary intermediate entries on the history stack. If we
don't do this, navigating back to the page before the redirection is
super hard and causes infinite loops sometimes. Also the refactoring
makes the code flow slightly easier to read.