Kuberay wrapper plugin (#467)
We will need a way to convert between a "user-oriented" way to specify a
Ray cluster (python dataclass object, describing in terms of cpu counts,
memory, gpu information), and an "infra oriented" way (a Kuberay K8s CRD
for a RayCluster). This PR provides an implementation of a conversion
between these two representations. Because there are a wide variety of
ways this conversion could be done, a wide variety of ways to configure
the RayCluster CRDs, and even some instability in the structure of the
CRDs between different Kuberay versions, this conversion is done via a
server-side plugin. A default plugin implementation is provided, which
should cover people using the most recent Kuberay and a conventional K8s
setup for GPU access.

With respect to GPU access, the default implementation provided will
support GPU usage via plugin configurations. All the plugin
configurations should be provided as json encoded into strings. The
configurations include:
- turning on GPU support (boolean flag)
- allowing specifying different tolerations for Ray workers/head nodes
that leverage GPUs
- allowing specifying different node selectors for Ray workers/head
nodes that leverage GPUs
- allowing specifying a key to use in resource requests/limits for GPU
count

node selectors and tolerations are also configurable for non-GPU
workers/heads. We know that existing users provide GPU support via
tolerations & node selectors.