Update Helm deployment instructions (#490)
Updated docs on Helm installation instructions. One item that is still a
TODO here (at the very end of the diff) is to create a table in a
`README.md` in the `helm-charts` repository that contains a list of all
of the possible settings, defaults, and explanations possible in
`values.yaml`.