Implements UI for external resource (#472)
This PR works to display the external resources tab. 

Currently, after the tab is switched to, the code will constantly poll
the server with an interval of 1 second until it sees the external
resource with `DEACTIVATED` state. Navigating away from the external
resources tab will cancel polling.

Despite the server returning a list of resources, we assume the server
will return a list of 1.

---------

Co-authored-by: tscurtu <tudor@sematic.dev>