Make Kuberay wrapper configurable via Helm chart (#525)
In order to use Ray with GPUs, we need to configure how the server
interprets requests for GPU, on the server side. This PR exposes the
Kuberay configurations in the Helm chart.

Testing
-------
Deployed the server to the dev environment with a variety of settings
for GPU/non GPU enabled Ray workers. I launched a variety of Ray cluster
jobs with these settings at different values, and examined the pods that
were created for the Ray workers to ensure they had the right
tolerations, node selectors, and requests.

---------

Co-authored-by: Sash Nagarkar <sashn45@gmail.com>
Co-authored-by: tscurtu <tudor@sematic.dev>