Improve run/resolution cleanup (clearer root cause, resolver restarts) (#531)
Addresses #518, #417

Make some improvements to the way run/resolution cleanup is performed.
In particular:
- Make it so that when one run fails and others are running/pending, it
is clear which have failed and which were just canceled
- Make it so when the resolver restarts, all runs are stopped (instead
of left running). We also add an appropriate error visible in the UI
when this occurs.

The first of these is achieved simply by changing the existing logic
that marks remaining runs as "failed" to mark them as "canceled"
instead. The latter is achieved by adjusting the logic such that when a
resolution reaches a terminal state, we always cancel any remaining
associated runs on the server side. This works because when a resolver
pod restarted, we were already failing the resolution. The only reason
the existing cleanup logic didn't cancel the existing runs was that
after a restart, the resolver no longer has in-memory representations of
the remote runs it had generated on its previous execution.

Testing
--------
In addition to the unit tests shown here, I also emulated the two cases
being addressed.

For emulating a resolver pod restart, I eliminated the SIGTERM handler
in my resolver code, then deleted the resolver pod with `kubectl` while
the resolution was in-progress. K8s automatically re-creates the pod in
this case, since there is still an associated k8s job. A resolution
created in this way is
[here](https://dev1.dev-usw2-sematic0.sematic.cloud/pipelines/sematic.examples.testing_pipeline.pipeline.testing_pipeline/4b3fd510027f45fe969eeb43d13aa9cf)
<img width="1138" alt="Screen Shot 2023-02-06 at 6 48 03 AM"
src="https://user-images.githubusercontent.com/7181589/217003187-a61c116b-019d-402d-a7c9-6ebeb0ba0272.png">

For emulating a run failure mid-pipeline, I simply used the testing
pipeline with fan-out and forced one of the fanned-out runs to fail with
an intentional exception. An example created in this way is
[here](https://dev1.dev-usw2-sematic0.sematic.cloud/pipelines/sematic.examples.testing_pipeline.pipeline.testing_pipeline/bc3450afe64341e4aff952c3a0569678)

<img width="1157" alt="Screen Shot 2023-02-06 at 6 49 36 AM"
src="https://user-images.githubusercontent.com/7181589/217003548-ed987172-a7a2-4403-9657-bcae0a9fbcdc.png">

---------

Co-authored-by: Josh Bauer <josh@sematic.dev>